requirements.txt
# 核心依赖
transformers>=4.35.0
torch>=2.0.0
peft>=0.6.0
bitsandbytes>=0.41.0
accelerate>=0.24.0

# 数据处理
datasets>=2.14.0
pyyaml>=6.0

# 微信接入
itchat>=1.3.10

# 记忆系统（可选）
chromadb>=0.4.0

# 工具库
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0

# 日志
loguru>=0.7.0